+++ src/main/java/com/revolvingmadness/sculk/language/errors/ValueError.java
@@ -1,0 +1,7 @@
+package com.revolvingmadness.sculk.language.errors;
+
+public class ValueError extends Error {
+    public ValueError(String message) {
+        super(message);
+    }
+}